modifying verifier
